+++ front-end/src/api/axiosAPI.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 const pathAPI = {
   "localhost": "http://localhost:8080", 
-  "pc-lab": "http://172.19.190.190:8080"
+  "pc-lab": "http://172.19.190.132:8080"
 }
 const api = axios.create({
   baseURL: pathAPI["pc-lab"],
