+++ front-end/src/api/status.config.json
@@ -1,13 +1,23 @@
 {
-    "userStatus": {
-        "login": {
-            "406" : {"message" : "user logged"}
-        },
-        "logout": {
-            
-        },
-        "create": {
-            "200" : {"message" : "Register sucess"}
+    "sectionStatus" : {
+        "userStatus": {
+            "login": {
+                "406" : {"message" : "user logged"},
+                "404" : {"message" : "user not found"}
+            },
+            "logout": {
+                
+            },
+            "register": {
+                "200" : {"message" : "Register sucess"}
+            }
+        }
+    },
+
+    "generalStatus" : {
+        "serveErrors" : {
+            "500" : {"message" : "internal server error"},
+            "ECONNABORTED" : {"message" : "service unavailable"}
         }
     }
 }
