+++ front-end/src/firebase.config.example.json
