+++ front-end/src/main.js
@@ -1,17 +1,18 @@
 import Vue from 'vue'
+import Vuex from 'vuex'
+
 import App from './views/App.vue'
 import routes from './routes.js'
 import VueRouter from 'vue-router';
 import firebase from  'firebase';
-import userStore from "./store/userStore.js"
-import Vuex from "vuex";
+import mainStore from './store/index.js'
+
 import {config} from './firebase.config.json';
 
-
-Vue.use(VueRouter)
 Vue.use(Vuex)
+Vue.use(VueRouter)
 
-const store = new Vuex.Store(userStore);
+const store = new Vuex.Store(mainStore);
 const router = new VueRouter({
   routes
 });
