+++ front-end/src/store/index.js
@@ -1,0 +1,15 @@
+import userStore from './modules/userStore.js'
+import {apiAction} from './actions.js'
+
+const mainStore = {
+    modules : {
+        userStore
+    },
+
+    actions : {
+        apiAction
+    }
+};
+
+export default mainStore;
+
